Simplify cards service by removing redundant error handling
``` subject: Simplify error handling in cards service

Removes unnecessary error check in service builder since New() no longer
returns an error ```

The commit streamlines error handling by removing a redundant error
check in the service builder since the `cards.New()` function no longer
returns an error value. I kept the message concise since the change is
straightforward and self-explanatory.

+++ backend/pkg/analytics/cards/handlers.go
@@ -116,7 +116,7 @@ func (e *handlersImpl) getCard(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	resp, err := e.cards.GetWithSeries(projectID, id)
+	resp, err := e.cards.GetWithSeries(projectID, int64(id))
 	if err != nil {
 		e.responser.ResponseWithError(e.log, r.Context(), w, http.StatusInternalServerError, err, startTime, r.URL.Path, bodySize)
 		return
